% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter.R
\name{particle_filter}
\alias{particle_filter}
\title{Compute posterior mean of the latent state (position and velocity).}
\usage{
particle_filter(y, dt, location1, location2, sr, q1, q2, statepriorMu,
  statepriorCholesky, importanceCholesky, nParticles)
}
\arguments{
\item{y}{A two-column matrix with the measurements.}

\item{dt}{The time step between observations.}

\item{location1}{A two-element vector with the longitude (x) and latitude
(y) of the first sensor.}

\item{location2}{A two-element vector with the longitude (x) and latitude
(y) of the second sensor.}

\item{sr}{The variance of the measurement model error.}

\item{q1}{The first difussion constant of the state model.}

\item{q2}{The second difussion constant of the state model.}

\item{statepriorMu}{A two-element vector with the longitude (x) and latitude
(y) of the location where the state prior density should be centered.}

\item{statepriorCholesky}{A four-element vector with the diagonal of the
Cholesky factor corresponding to the variance of the state prior distribution.}

\item{importanceCholesky}{A four-element vector with the diagonal of the
Cholesky factor corresponding to the variance of the importance distribution.}

\item{nParticles}{An integer with the number of particles.}
}
\value{
A named list with four elements.
`noiseless` is a T x 2 matrix with the noiseless approximation of the
vehicle position (assumes no noise and velocity equal to zero).
`stateMean` is a T x 4 matrix with the posterior mean of the latent state
at each time step.
`weights` is a T x nParticles matrix with the normalized weights.
`ess` is a T-sized vector with the effective sample size at each time step.
}
\description{
For a given parameter vector, this function runs a Particle Filter to
estimate the posterior mean of the latent state for the bearing-only
tracking problem with two passive sensors.
}
\note{
The resampling step is currently not implemented. Expect particle
degeneracy (i.e. rapidly decaying ESS).
}
\seealso{
\code{\link{plot.filtered}{plot}}
}
