% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter.R
\name{plot.filtered}
\alias{plot.filtered}
\title{Plot the posterior mean of the position as estimated by the Particle Filter.}
\usage{
\method{plot}{filtered}(x, ...)
}
\arguments{
\item{x}{An object returned by the function \code{\link{particle_filter}}.}

\item{...}{Currently nothing.}
}
\description{
Plot the posterior mean of the position as estimated by the Particle Filter.
}
\note{
The function may hide values of the state posterior mean that are
estimated as zero (funny estimates).
}
