% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter.R
\docType{data}
\name{vehicle}
\alias{vehicle}
\title{Measurements from two passive sensors tracking a moving vehicle.}
\format{A data frame with 11027 rows and 2 numeric variables:
\describe{
  \item{a1}{Angle in radians of a moving vehicle as measured from one
  passive sensor located at -93.2494663765932, 41.5563518606521.}
  \item{a2}{Angle in radians of a moving vehicle as measured from another
  passive sensor located at -93.2475338232000, 41.5576632356000.}
}}
\source{
\url{https://github.com/ISU-STRIPS/STRIPSyield/}
}
\usage{
vehicle
}
\description{
Measurements from two passive sensors tracking a moving vehicle.
}
\keyword{datasets}
